\name{excel_sheets}
\alias{excel_sheets}
\title{List sheet names in an Excel workbook}
\usage{excel_sheets(path)}
\arguments{\item{path}{Path to an Excel workbook, or a raw vector containing workbook bytes.}}
\value{A character vector of sheet names.}
\description{Lists sheet names in an Excel workbook.}
