\name{excel_tables}
\alias{excel_tables}
\title{List table names in an Excel workbook}
\usage{excel_tables(path, sheet = NULL)}
\arguments{
\item{path}{Path to an Excel workbook.}
\item{sheet}{Optional sheet name used to limit results to one worksheet.}
}
\value{A character vector of table names.}
\description{Lists table names in an Excel workbook.}
