% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{excel_sheets}
\alias{excel_sheets}
\title{List sheet names in an Excel workbook}
\usage{
excel_sheets(file)
}
\arguments{
\item{file}{Path to an Excel workbook, or a raw vector containing workbook
bytes.}
}
\value{
A character vector of sheet names.
}
\description{
List sheet names in an Excel workbook
}
