\docType{package}
\name{fastexcel-package}
\alias{fastexcel}
\alias{fastexcel-package}
\title{fastexcel: Fast Excel Reader Backed by Rust}
\description{Reads Excel workbooks using the Rust \code{fastexcel} crate from ToucanToco and returns Arrow-first results in R.}
\details{The upstream Rust crate is MIT licensed and lives at \url{https://github.com/ToucanToco/fastexcel}.}
\keyword{internal}
